requirements.txt
astroid==2.0.4
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.0
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
httpie==1.0.2
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.6.1
Pygments==2.3.1
pylint==2.1.1
pylint-django==2.0.4
pylint-plugin-utils==0.4
pytz==2018.7
requests==2.21.0
six==1.11.0
urllib3==1.24.1
virtualenv==16.1.0
wrapt==1.10.11
